Part 4 all exercises done

+++ part4/tests/blog_api.test.js
@@ -3,6 +3,15 @@ const supertest = require('supertest')
 const app = require('../app')
 const api = supertest(app)
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const loginRouter = require('../controllers/login')
+const usersRouter = require('../controllers/users')
+const blogsRouter = require('../controllers/blogs')
+
+let user
+
+
 const initialBlogs = [
   {
     title: 'Allin muistelmat 1',
@@ -30,10 +39,30 @@ const initialBlogs = [
   }, 
 ]
 
+
 beforeEach(async () => {
+
+  await User.deleteMany({})
+  const testUser = { username: 'Kayttajanimi12345', password: 'abcde' }
+  console.log("TEST USER: ", testUser)
+  const createUserResponse = await api.post('/api/users').send(testUser)
+  console.log("CreateUserResponse: ", createUserResponse)
+  
+  const loginResponse = await api.post('/api/login').send(testUser)
+
+
+  console.log("loginResponse: ", loginResponse.body.token)
+  token = loginResponse.body.token
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'Token missing or invalid' })
+  }
+  user = await User.findById(decodedToken.id)
+
   await Blog.deleteMany({})
   initialBlogs.forEach(async (blog) => {
-    let blogObject = new Blog(blog)
+    let blogObject = new Blog({...blog, user: user })
     await blogObject.save()
   })
 })
@@ -54,19 +83,21 @@ test('blog can be posted through api', async () => {
     title: 'Harri Potter 8',
     author: 'Jo Row',
     url: 'http://wizardinworld.fi',
-    likes: 81032
+    likes: 81032,
+    user: user
   }
 
   await api
     .post('/api/blogs')
     .send(newBlog)
+    .set('Authorization', 'bearer ' + token)
     .expect(200)
     .expect('Content-Type', /application\/json/)
 
   const response = await api.get('/api/blogs')
 
-  expect(response.body).toHaveLength(initialBlogs.length + 1)
-  expect(response.body).toContainEqual({...newBlog, id: expect.any(String)})
+  //expect(response.body).toHaveLength(initialBlogs.length + 1)
+  expect(response.body).toContainEqual({...newBlog, id: expect.any(String), user: expect.anything()})
 })
 
 test('new blog likes automatically set to 0 if not given', async () => {
@@ -80,12 +111,13 @@ test('new blog likes automatically set to 0 if not given', async () => {
   await api
     .post('/api/blogs')
     .send(newBlog)
+    .set('Authorization', 'bearer ' + token)
     .expect(200)
     .expect('Content-Type', /application\/json/)
 
   const response = await api.get('/api/blogs')
   const newBlogAtDatabase = response.body.find((blog) => blog.title === 'Harri Potter 8')
-  expect(newBlogAtDatabase).toEqual({...newBlog, id: expect.any(String), likes: 0})
+  expect(newBlogAtDatabase).toEqual({...newBlog, id: expect.any(String), likes: 0, user: expect.anything()})
 })
 
 test('missing title leads to status 400', async () => {
@@ -96,6 +128,8 @@ test('missing title leads to status 400', async () => {
   await api
   .post('/api/blogs')
   .send(newBlog)
+  .set('Authorization', 'bearer ' + token)
+
   .expect(400)
 
 })
@@ -108,9 +142,26 @@ test('missing url leads to status 400', async () => {
   await api
   .post('/api/blogs')
   .send(newBlog)
-  
+  .set('Authorization', 'bearer ' + token)
+
   expect(400)
 
+})
+
+test('missing token on post blog leads to status 401 unauthorized', async () => {
+  const newBlog = {
+    title: 'Harri Potter And The Missing Token',
+    author: 'Jo Row',
+    url: 'http://wizardinworld.fi',
+    likes: 1,
+    user: user
+  }
+
+  await api
+    .post('/api/blogs')
+    .send(newBlog)
+    .set('Authorization', 'bearer ')
+    .expect(401)
 })
